requirements.txt
# Core game dependencies
pygame==2.5.2
PyOpenGL==3.1.7
torch==2.0.1
stable-baselines3[extra]>=2.0.0a5


# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Optional GUI
tkinter-tooltip==2.1.0